Moves url parser function to a better location. Removes URL validation.
